Resync with latest changes in GenomicFeatures
supportedGeneIDs() and downloadLengthFromUCSC() were broken because
of recent changes in GenomicFeatures. This is a quick-and-dirty fix
so the geneLenDataBase and goseq packages can build and pass check
again on the build system. Note that with this fix downloadLengthFromUCSC()
doesn't support id="sgdGene" but maybe it isn't aimed at supporting it
(right now the package only seems to be targetting id="ensGene" for
Yeast). Anyway maybe a better fix for downloadLengthFromUCSC()
would be to use the functionality provided by
GenomicFeatures:::.fetchTxName2GeneIdMappingFromUCSC()
rather than trying to reimplement it.


git-svn-id: file:///home/git/bioc-data.hedgehog.fhcrc.org/trunk/experiment/pkgs/geneLenDataBase@1210 db2202da-8704-0410-b924-b125b3cd84ef